requirements.txt
--index-url https://pypi.python.org/simple/


pytest-cov